requirements.txt
dgg-bot
websocket-client
google-cloud-storage
schedule